Update DirectMessageForm and DirectMessageList components
Added a Select component to DirectMessageForm to allow selecting a recipient from a list of users.
Fetch users every 3 seconds in DirectMessageForm and store them in the state.
Updated DirectMessageList to display messages grouped by author.
Added a button to expand messages from a specific author in DirectMessageList.
Commented out the scroll functionality in Messages component as it was causing issues.
Made some minor UI changes in Home component.
Updated getDirectMessages controller in the server to use the receiverId from the request.
Modified createDirectMessage controller in the server to include the userName of the author in the message.
Added a comment to include a script for updating all messages with a userName field in the database.